fix #14 IHL unclear
The Internet Header Length is expressed in 32bits words; previously we
were automagically multiplying IHL by 4 to return a bytes length, but as
discussed in issue #14 this was probably a bad idea so now we just
return it straight.
NB: maybe this should have been a breaking change, if it causes
some problem please report, I'll yank this version and publish a sermver
change.